Add main ruby and data file for msfvc
Ruby program resides in the plugins directory and the data json file
is in the data/wordlists directory.

Add welcome message when msfvc plugin is loade.

Add base plugin metasploit-framework integration to msfvc.rb with
required method definitions and an ascii art welcome message.

Add a CommandDispatcher with two commands for the console.

- vc_list will list available voice commands
- vc_help will display a help message for commands