Add vc_details command and seach to vc_list
vc_details command give details on a voice command. The voice command
is selected via a search or a specific id for a voice assistant.

Add search to vc_list for a search of all fields of voice commands and
verbose mode. Verbose shows all fields. Only id, command, and category
are shown without verbose specified.