Refactor VoiceCmd to take fill arguments.
Refactor add_to_script to use VoiceCmd fill arguments.

Reorder option execution in vc_script to allow more options to be used
together.

Alter vc_script remove option to use an index rather than assistant
and id allowing for the case when multiple of the same command are in
a script.